feat: 이미지 캐러셀 및 색상칩 컴포넌트 분리
